+++ TESTING/LIN/stsqr01.f
@@ -210,7 +210,7 @@
       ANORM = SLANGE( '1', M, N, A, M, RWORK )
       RESID = SLANGE( '1', M, N, R, M, RWORK )
       IF( ANORM.GT.ZERO ) THEN
-         RESULT( 1 ) = RESID / (EPS*MAX(1,M)*ANORM)
+         RESULT( 1 ) = RESID / (EPS*REAL( MAX( 1, M ) )*ANORM)
       ELSE
          RESULT( 1 ) = ZERO
       END IF
@@ -220,7 +220,7 @@
       CALL SLASET( 'Full', M, M, ZERO, ONE, R, M )
       CALL SSYRK( 'U', 'C', M, M, -ONE, Q, M, ONE, R, M )
       RESID = SLANSY( '1', 'Upper', M, R, M, RWORK )
-      RESULT( 2 ) = RESID / (EPS*MAX(1,M))
+      RESULT( 2 ) = RESID / (EPS*REAL( MAX( 1, M ) ))
 *
 *     Generate random m-by-n matrix C and a copy CF
 *
@@ -241,7 +241,7 @@
       CALL SGEMM( 'N', 'N', M, N, M, -ONE, Q, M, C, M, ONE, CF, M )
       RESID = SLANGE( '1', M, N, CF, M, RWORK )
       IF( CNORM.GT.ZERO ) THEN
-         RESULT( 3 ) = RESID / (EPS*MAX(1,M)*CNORM)
+         RESULT( 3 ) = RESID / (EPS*REAL( MAX( 1, M ) )*CNORM)
       ELSE
          RESULT( 3 ) = ZERO
       END IF
@@ -261,7 +261,7 @@
       CALL SGEMM( 'T', 'N', M, N, M, -ONE, Q, M, C, M, ONE, CF, M )
       RESID = SLANGE( '1', M, N, CF, M, RWORK )
       IF( CNORM.GT.ZERO ) THEN
-         RESULT( 4 ) = RESID / (EPS*MAX(1,M)*CNORM)
+         RESULT( 4 ) = RESID / (EPS*REAL( MAX( 1, M ) )*CNORM)
       ELSE
          RESULT( 4 ) = ZERO
       END IF
@@ -285,7 +285,7 @@
       CALL SGEMM( 'N', 'N', N, M, M, -ONE, D, N, Q, M, ONE, DF, N )
       RESID = SLANGE( '1', N, M, DF, N, RWORK )
       IF( DNORM.GT.ZERO ) THEN
-         RESULT( 5 ) = RESID / (EPS*MAX(1,M)*DNORM)
+         RESULT( 5 ) = RESID / (EPS*REAL( MAX( 1, M ) )*DNORM)
       ELSE
          RESULT( 5 ) = ZERO
       END IF
@@ -304,7 +304,7 @@
       CALL SGEMM( 'N', 'T', N, M, M, -ONE, D, N, Q, M, ONE, DF, N )
       RESID = SLANGE( '1', N, M, DF, N, RWORK )
       IF( CNORM.GT.ZERO ) THEN
-         RESULT( 6 ) = RESID / (EPS*MAX(1,M)*DNORM)
+         RESULT( 6 ) = RESID / (EPS*REAL( MAX( 1, M ) )*DNORM)
       ELSE
          RESULT( 6 ) = ZERO
       END IF
@@ -354,7 +354,7 @@
       ANORM = SLANGE( '1', M, N, A, M, RWORK )
       RESID = SLANGE( '1', M, N, LQ, L, RWORK )
       IF( ANORM.GT.ZERO ) THEN
-         RESULT( 1 ) = RESID / (EPS*MAX(1,N)*ANORM)
+         RESULT( 1 ) = RESID / (EPS*REAL( MAX( 1, N ) )*ANORM)
       ELSE
          RESULT( 1 ) = ZERO
       END IF
@@ -364,7 +364,7 @@
       CALL SLASET( 'Full', N, N, ZERO, ONE, LQ, L )
       CALL SSYRK( 'U', 'C', N, N, -ONE, Q, N, ONE, LQ, L )
       RESID = SLANSY( '1', 'Upper', N, LQ, L, RWORK )
-      RESULT( 2 ) = RESID / (EPS*MAX(1,N))
+      RESULT( 2 ) = RESID / (EPS*REAL( MAX( 1, N ) ))
 *
 *     Generate random m-by-n matrix C and a copy CF
 *
@@ -384,7 +384,7 @@
       CALL SGEMM( 'N', 'N', N, M, N, -ONE, Q, N, D, N, ONE, DF, N )
       RESID = SLANGE( '1', N, M, DF, N, RWORK )
       IF( DNORM.GT.ZERO ) THEN
-         RESULT( 3 ) = RESID / (EPS*MAX(1,N)*DNORM)
+         RESULT( 3 ) = RESID / (EPS*REAL( MAX( 1, N ) )*DNORM)
       ELSE
          RESULT( 3 ) = ZERO
       END IF
@@ -403,7 +403,7 @@
       CALL SGEMM( 'T', 'N', N, M, N, -ONE, Q, N, D, N, ONE, DF, N )
       RESID = SLANGE( '1', N, M, DF, N, RWORK )
       IF( DNORM.GT.ZERO ) THEN
-         RESULT( 4 ) = RESID / (EPS*MAX(1,N)*DNORM)
+         RESULT( 4 ) = RESID / (EPS*REAL( MAX( 1, N ) )*DNORM)
       ELSE
          RESULT( 4 ) = ZERO
       END IF
@@ -426,7 +426,7 @@
       CALL SGEMM( 'N', 'N', M, N, N, -ONE, C, M, Q, N, ONE, CF, M )
       RESID = SLANGE( '1', N, M, DF, N, RWORK )
       IF( CNORM.GT.ZERO ) THEN
-         RESULT( 5 ) = RESID / (EPS*MAX(1,N)*CNORM)
+         RESULT( 5 ) = RESID / (EPS*REAL( MAX( 1, N ) )*CNORM)
       ELSE
          RESULT( 5 ) = ZERO
       END IF
@@ -445,7 +445,7 @@
       CALL SGEMM( 'N', 'T', M, N, N, -ONE, C, M, Q, N, ONE, CF, M )
       RESID = SLANGE( '1', M, N, CF, M, RWORK )
       IF( CNORM.GT.ZERO ) THEN
-         RESULT( 6 ) = RESID / (EPS*MAX(1,N)*CNORM)
+         RESULT( 6 ) = RESID / (EPS*REAL( MAX( 1, N ) )*CNORM)
       ELSE
          RESULT( 6 ) = ZERO
       END IF
